+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file is used to list changes made in each version of the aws-parallelcluste
 
 **CHANGES**
 - Use inclusive language in internal naming convention.
+- Improve error handling in slurm plugin processes when clustermgtd is down.
 
 2.10.0
 -----